requirements.txt
pyeda==0.28.0
python-statemachine==0.8.0
